+++ tests/plugin/pinning/test.php
@@ -1,0 +1,7 @@
+<?php
+
+use Http\Discovery\Psr17FactoryDiscovery;
+
+require __DIR__.'/vendor/autoload.php';
+
+echo get_class(Psr17FactoryDiscovery::findRequestFactory())."\n";